chore: Renaming/consolidating files for clarity
