+++ src/formatter.ts
@@ -1,214 +1,139 @@
-import { blueBright, dim, red, reset, underline, yellow } from 'chalk';
-import type { ESLint } from 'eslint';
-import stripAnsi from 'strip-ansi';
-import table from 'text-table';
+import {
+  applyTodoChanges,
+  buildTodoData,
+  getTodoBatches,
+  getTodoStorageDirPath,
+  readTodos,
+  TodoData,
+  todoFilePathFor,
+  todoStorageDirExists,
+  writeTodos,
+  _buildTodoDatum,
+} from '@ember-template-lint/todo-utils';
+import type { ESLint, Linter } from 'eslint';
+import { format } from './format';
+import { TodoFormatterOptions, TodoResultMessage } from './types';
+import { getBaseDir } from './get-base-dir';
+import hasFlag from 'has-flag';
 import { ERROR_SEVERITY, TODO_SEVERITY } from './constants';
-import type {
-  TodoFormatterCounts,
-  TodoFormatterOptions,
-  TodoResultMessage,
-} from './types';
 
-export function formatter(
-  results: ESLint.LintResult[],
-  { includeTodo }: TodoFormatterOptions = {
-    includeTodo: false,
+export function formatter(results: ESLint.LintResult[]): void {
+  // note: using async/await directly causes eslint to print `Promise { <pending> }`
+  formatterAsync(results);
+}
+
+export async function formatterAsync(results: ESLint.LintResult[]): Promise<void> {
+  const updateTodo = process.env.UPDATE_TODO === '1';
+  const includeTodo = process.env.INCLUDE_TODO === '1';
+
+  if (updateTodo) {
+    await writeTodos(getBaseDir(), results);
   }
-): string {
-  const counts = tallyResults(results);
 
-  let output = '\n';
-
-  output += formatResults(results, { includeTodo });
-
-  output += formatSummary(counts, { includeTodo });
-
-  // Resets output color to prevent change on top level
-  return counts.total > 0 || (includeTodo && counts.todoCount > 0)
-    ? reset(output)
-    : '';
+  await report(results, { includeTodo });
 }
 
-function formatResults(
+/**
+ * Mutates all errors present in the todo dir to todos in the results array.
+ *
+ * @param results ESLint results array
+ */
+export async function transformResults(
+  baseDir: string,
   results: ESLint.LintResult[],
-  options: TodoFormatterOptions
-): string {
-  let output = '';
-
+  todoMap: Map<string, TodoData>
+): Promise<void> {
   results.forEach((result) => {
-    const messages = result.messages as TodoResultMessage[];
-
-    if (messages.length === 0) {
-      return;
-    }
-
-    const areAllMessagesTodo = messages.every(
-      (message) => message.severity === TODO_SEVERITY
-    );
-
-    if (options.includeTodo || !areAllMessagesTodo) {
-      output += `${underline(result.filePath)}\n`;
-    }
-
-    output += `${formatMessages(messages, options)}`;
-    output += !options.includeTodo && areAllMessagesTodo ? '' : '\n\n';
-  });
-
-  return output;
-}
-
-function formatMessages(
-  messages: TodoResultMessage[],
-  options: TodoFormatterOptions
-): string {
-  const messageRows = messages
-    .filter(
-      (message) => message.severity !== TODO_SEVERITY || options.includeTodo
-    )
-    .map((message) => {
-      let messageType;
-
-      if (message.fatal || message.severity === ERROR_SEVERITY) {
-        messageType = red('error');
-      } else if (message.severity === TODO_SEVERITY) {
-        messageType = blueBright('todo');
-      } else {
-        messageType = yellow('warning');
+    (result.messages as TodoResultMessage[]).forEach((message) => {
+      if (message.severity !== ERROR_SEVERITY) {
+        return;
       }
 
-      return [
-        '',
-        message.line || 0,
-        message.column || 0,
-        messageType,
-        message.message.replace(/([^ ])\.$/u, '$1'),
-        dim(message.ruleId || ''),
-      ];
+      // we only mutate errors that are present in the todo map, so check if it's there first
+      const todoDatum = _buildTodoDatum(
+        baseDir,
+        result,
+        message as Linter.LintMessage
+      );
+
+      const todoHash = todoFilePathFor(todoDatum);
+
+      if (!todoMap.has(todoHash)) {
+        return;
+      }
+
+      message.severity = TODO_SEVERITY;
+
+      result.errorCount -= 1;
+      result.todoCount = Number.isInteger(result.todoCount)
+        ? result.todoCount + 1
+        : 1;
+
+      if (!message.fix) {
+        return;
+      }
+
+      result.fixableErrorCount -= 1;
+      result.fixableTodoCount = Number.isInteger(result.fixableTodoCount)
+        ? result.fixableTodoCount + 1
+        : 1;
     });
-
-  const messageTableOptions: table.Options = {
-    align: ['.', 'r', 'l'],
-    stringLength(str) {
-      return stripAnsi(str).length;
-    },
-  };
-
-  return messageRows.length > 0
-    ? table(messageRows, messageTableOptions)
-        .split('\n')
-        .map((el) =>
-          el.replace(/(\d+)\s+(\d+)/u, (m, p1, p2) => dim(`${p1}:${p2}`))
-        )
-        .join('\n')
-    : '';
+  });
 }
 
-function formatSummary(
-  counts: TodoFormatterCounts,
+async function report(
+  results: ESLint.LintResult[],
   options: TodoFormatterOptions
 ) {
-  let output = '';
+  const baseDir = getBaseDir();
 
-  const { includeTodo } = options;
+  if (todoStorageDirExists(baseDir)) {
+    const existingTodoFiles = await readTodos(baseDir);
+    const [, itemsToRemoveFromTodos,] = await getTodoBatches(
+      buildTodoData(baseDir, results),
+      existingTodoFiles
+    );
 
-  const {
-    total,
-    errorCount,
-    warningCount,
-    todoCount,
-    fixableErrorCount,
-    fixableWarningCount,
-    fixableTodoCount,
-  } = counts;
-
-  if (total > 0 || (includeTodo && todoCount > 0)) {
-    const chalkColorFunction = errorCount > 0 ? red : yellow;
-
-    let summary = [
-      '\u2716 ',
-      total,
-      pluralize(' problem', total),
-      ' (',
-      errorCount,
-      pluralize(' error', errorCount),
-      ', ',
-      warningCount,
-      pluralize(' warning', warningCount),
-    ];
-
-    if (includeTodo) {
-      summary = [...summary, ', ', todoCount, pluralize(' todo', todoCount)];
+    if (itemsToRemoveFromTodos.size > 0) {
+      if (hasFlag('fix')) {
+        applyTodoChanges(getTodoStorageDirPath(baseDir), new Map(), itemsToRemoveFromTodos);
+      } else {
+        itemsToRemoveFromTodos.forEach((todo) => {
+          pushResult(results, todo);
+        });
+      }
     }
 
-    summary = [...summary, ')\n'];
-
-    output += chalkColorFunction.bold(summary.join(''));
-
-    if (
-      fixableErrorCount > 0 ||
-      fixableWarningCount > 0 ||
-      (includeTodo && fixableTodoCount > 0)
-    ) {
-      let fixableMessage = includeTodo
-        ? [
-            '  ',
-            fixableErrorCount,
-            pluralize(' error', fixableErrorCount),
-            ', ',
-            fixableWarningCount,
-            pluralize(' warning', fixableWarningCount),
-            ', and ',
-            fixableTodoCount,
-            pluralize(' todo', fixableTodoCount),
-          ]
-        : [
-            '  ',
-            fixableErrorCount,
-            pluralize(' error', fixableErrorCount),
-            ' and ',
-            fixableWarningCount,
-            pluralize(' warning', fixableWarningCount),
-          ];
-
-      fixableMessage = [
-        ...fixableMessage,
-        ' potentially fixable with the `--fix` option.\n',
-      ];
-
-      output += chalkColorFunction.bold(fixableMessage.join(''));
-    }
-
-    output += '\n';
+    await transformResults(baseDir, results, existingTodoFiles);
   }
 
-  return output;
+  process.stdout.write(format(results, options));
 }
 
-function tallyResults(results: ESLint.LintResult[]): TodoFormatterCounts {
-  const counts = {
-    total: 0,
-    errorCount: 0,
-    warningCount: 0,
-    todoCount: 0,
-    fixableErrorCount: 0,
-    fixableWarningCount: 0,
-    fixableTodoCount: 0,
-  };
+function pushResult(results: ESLint.LintResult[], todo: TodoData) {
+  const resultForFile = results.find((r: ESLint.LintResult) => r.filePath === todo.filePath);
+  const result: Linter.LintMessage = {
+    ruleId: 'invalid-todo-violation-rule',
+    message: `Todo violation passes \`${todo.ruleId}\` rule. Please run \`--fix\` to remove this todo from the todo list.`,
+    severity: 2,
+    column: 0,
+    line: 0,
+  }
 
-  results.forEach((result) => {
-    counts.errorCount += result.errorCount;
-    counts.warningCount += result.warningCount;
-    counts.todoCount += result.todoCount || 0;
-    counts.fixableErrorCount += result.fixableErrorCount;
-    counts.fixableWarningCount += result.fixableWarningCount;
-    counts.fixableTodoCount += result.fixableTodoCount || 0;
-  });
-
-  counts.total = counts.errorCount + counts.warningCount;
-
-  return counts;
-}
-
-function pluralize(word: string, count: number): string {
-  return count === 1 ? word : `${word}s`;
+  if (resultForFile) {
+    resultForFile.messages.push(result);
+    resultForFile.errorCount += 1;
+  } else {
+    results.push({
+      filePath: todo.filePath,
+      messages: [result],
+      errorCount: 1,
+      warningCount: 0,
+      todoCount: 0,
+      fixableErrorCount: 0,
+      fixableWarningCount: 0,
+      fixableTodoCount: 0,
+      usedDeprecatedRules: [],
+    });
+  }
 }
